- Added shared files to common section - Changed include statements to point to shared files - Setup the checksum code as a library - Modified the platformio.ini to take into account the new file locations
