+++ common/libs/checksum/checksum.cpp
@@ -1,0 +1,55 @@
+/* Copyright 2020, Edwin Chiu
+
+  This file is part of FixMoreLungs.
+
+  FixMoreLungs is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  FixMoreLungs is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with FixMoreLungs.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+#include <stdint.h>
+
+#include "checksum.h"
+
+uint16_t checksum_fletcher16(uint16_t *sum1, uint16_t *sum2,  char *data, uint8_t count)
+{
+    uint8_t index;
+
+    for (index = 0; index < count; ++index)
+    {
+        *sum1 = ((*sum1) + data[index]) % 255;
+        *sum2 = ((*sum2) + (*sum1)) % 255;
+    }
+
+    return ((*sum2) << 8) | *sum1;
+}
+
+bool checksum_check(char *packet, uint8_t packet_len) {
+    uint16_t sum1 = 0;
+    uint16_t sum2 = 0;
+    uint16_t csum;
+    uint8_t c0,c1,f0,f1;
+
+    // TODO Check validity of packet_len ?
+    // TODO Can this code be optimised?
+
+    csum = checksum_fletcher16(&sum1, &sum2, packet, packet_len-2); // Don't check the final two check bytes
+    f0 = csum & 0xff;
+    f1 = (csum >> 8) & 0xff;
+    c0 = 0xff - ((f0 + f1) % 0xff);
+    c1 = 0xff - ((f0 + c0) % 0xff);
+
+    if((packet[packet_len-2] == (char) c0) && (packet[packet_len-1] == (char) c1))
+        return true;
+
+    return false;
+}
