+++ controller/IDE/platformio/platformio.ini
@@ -11,7 +11,11 @@
 include_dir = controller/include
 src_dir = controller/src
 
+
 [env:uno]
 platform = atmelavr
 board = uno
 framework = arduino
+lib_ldf_mode = deep+
+lib_extra_dirs = common/libs/
+build_flags = -Icommon/include/
