+++ controller/include/comms.h
@@ -21,6 +21,9 @@
 
 #include <stdint.h>
 
+#include "checksum.h"
+
+#include "packet_types.h"
 
 enum class processPacket {
     checksumErr     = 0x00,
@@ -34,6 +37,10 @@ enum class processPacket {
     count                       /* Sentinel */
 };
 
+/* Used to store the minimum and maximum medical mode command numbers
+*  This data is used to validate the incoming commands to be sure that
+*  they're in the proper ranges.
+*/
 enum class medicalMode {
     start   = 0x00,     /* First medical mode command */
     end     = 0x10,     /* Final medical mode command */
@@ -41,6 +48,10 @@ enum class medicalMode {
     count
 };
 
+/* Used to store the minimum and maximum engineering mode command numbers
+*  This data is used to validate the incoming commands to be sure that
+*  they're in the proper ranges.
+*/
 enum class engMode {
     start   = 0x20,     /* First engineering mode command */
     end     = 0x27,     /* Final engineering mode command */
@@ -48,89 +59,15 @@ enum class engMode {
     count
 };
 
+/* Used to store the minimum and maximum mixed mode command numbers
+*  This data is used to validate the incoming commands to be sure that
+*  they're in the proper ranges.
+*/
 enum class mixedMode {
     start   = 0x40,     /* First mixed mode command */
     end     = 0x44,     /* Final mixed mode command */
 
     count
-};
-
-enum class command {
-
-    /* Medical mode commands */
-
-    set_rr          = 0x00,
-    get_rr          = 0x01,
-    set_tv          = 0x02,
-    get_tv          = 0x03,
-    set_peep        = 0x04,
-    get_peep        = 0x05,
-    set_pip         = 0x06,
-    get_pip         = 0x07,
-    set_dwell       = 0x08,
-    get_dwell       = 0x09,
-    set_id          = 0x0a,   /* Inspiration duration */
-    get_id          = 0x0b,
-    set_ed          = 0x0c,   /* Expiration duration */
-    get_ed          = 0x0d,
-
-    get_pressure    = 0x0e,
-    get_flow        = 0x0f,
-    get_volume      = 0x10,
-
-    /* Engineering mode commands */
-
-    set_kp          = 0x20,     /* PID constant Kp */
-    get_Kp          = 0x21,
-    set_Ki          = 0x22,     /* PID constant Ki */
-    get_Ki          = 0x23,
-    set_Kd          = 0x24,     /* PID constant Kd */
-    get_Kd          = 0x25,
-
-    set_blower      = 0x26,     /* Turn blower ON/OFF */
-    reset_vc        = 0x27,     /* Reset Ventilation Controller */
-
-    /* Mixed Engineering/Medical mode commands */
-
-    set_periodic    = 0x40,     /* Periodic data transmission mode (Pressure, Flow, Volume) */
-    get_periodic    = 0x41,
-    set_mode        = 0x42,     /* Engineering or medical mode */
-    get_mode        = 0x43,
-    comms_check     = 0x44,     /* Communications check command */
-
-    count                       /* Sentinel */
-};
-
-enum class msgType {
-    cmd             = 0x00,     /* Command */
-    ack             = 0x01,     /* VC alarm Ack */
-    nAck            = 0x02,     /* VC alarm Fail */
-
-    rAck            = 0x10,     /* Response Ack */
-    rErrChecksum    = 0x11,     /* Response checksum error */
-    rErrMode        = 0x12,     /* Response mode error */
-    rErrCmd         = 0x13,     /* Response cmd error */
-
-    status          = 0x20,     /* Status */
-    alarm           = 0x30,     /* Alarm */
-    data            = 0x40,     /* Data */
-
-    count                       /* Sentinel */
-};
-
-enum class dataID {
-    /* Status */
-    vc_boot      = 0x80,    /* Status sent when arduino boots (includes software version) */
-
-
-    /* Alarms */
-    alarm_1     = 0xA0,
-    alarm_2     = 0xA1,
-
-    /* Data */
-    data_1      = 0xC0,
-
-    count                   /* Sentinel */
 };
